requirements.txt
Django==3.0.6
psycopg2==2.8.5
django-unixtimestampfield==0.4.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0